anpr - Copy/requirements.txt
ultralytics==8.0.114
pandas==2.0.2
opencv-python==4.7.0.72
numpy==1.24.3
scipy==1.10.1
easyocr==1.7.0
filterpy==1.4.5
pyodbc==5.0.1
ultralytics
opencv-python
numpy
easyocr
google-auth-oauthlib
google-auth
google-api-python-client
Pillow>=7.2.0
